backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-JWT-Extended
python-dotenv
psycopg2-binary
pymysql
Werkzeug
flask_mail
requests
marshmallow
stripe